requirements.txt
absl-py==2.0.0
astor==0.8.1
certifi @ file:///croot/certifi_1671487769961/work/certifi
charset-normalizer==3.3.2
docopt==0.6.2
gast==0.2.2
google-pasta==0.2.0
grpcio==1.59.2
h5py==3.8.0
idna==3.4
importlib-metadata==6.7.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.4.4
MarkupSafe==2.1.3
neuralgym @ git+https://github.com/JiahuiYu/neuralgym@88292adb524186693a32404c0cfdc790426ea441
numpy==1.21.6
opencv-python==4.8.1.78
opt-einsum==3.3.0
Pillow==9.5.0
pipreqs==0.4.13
protobuf==3.19.0
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
tensorboard==1.15.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
termcolor==2.3.0
typing_extensions==4.7.1
urllib3==2.0.7
Werkzeug==2.2.3
wrapt==1.15.0
yarg==0.1.9
zipp==3.15.0
